Rw/meta (#139)
* initial setup

* unused file

* handle can now add syntax trees

* Added CompilerContext.GetSymbolsWithName

* bug fixing

* updated docs

* fixed directive width bug

* docs

+++ src/Buckle/Compiler/Libraries/Standard/StandardLibrary.cs
@@ -163,8 +163,8 @@ internal static class StandardLibrary {
 
     private static SynthesizedFinishedNamedTypeSymbol GenerateCallingConvention() {
         return StaticClass("CallingConvention", [
-            ConstExprField("Winapi", SpecialType.UInt32, 1),
-            ConstExprField("Cdecl", SpecialType.UInt32, 2),
+            ConstExprField("Winapi", SpecialType.UInt32, (uint)1),
+            ConstExprField("Cdecl", SpecialType.UInt32, (uint)2),
         ]);
     }
